requirements.txt
langchain
langchain_community
python-dotenv
streamlit
langchain_groq
langchain_huggingface
ipykernel
langchain_core
